refactor: follow changes of gtfs_parser
