--- gtfs_parser/README.md
@@ -1,5 +1,0 @@
-# gtfs_parser
-
-## LICENSE
-
-MIT License
