+++ gtfs_parser/__init__.py
@@ -1,1 +1,2 @@
-from .__main__ import GTFSParser
+from .gtfs import GTFS
+from . import aggregate, parse
