--- gtfs_parser/tests/fixture/fare_attributes.txt
@@ -1,107 +1,0 @@
-﻿fare_id,price,currency_type,payment_method,transfers
-100_00,100,JPY,0,0
-130_00,130,JPY,0,0
-140_00,140,JPY,0,0
-150_00,150,JPY,0,0
-160_00,160,JPY,0,0
-170_00,170,JPY,0,0
-180_00,180,JPY,0,0
-190_00,190,JPY,0,0
-200_00,200,JPY,0,0
-210_00,210,JPY,0,0
-220_00,220,JPY,0,0
-230_00,230,JPY,0,0
-240_00,240,JPY,0,0
-250_00,250,JPY,0,0
-260_00,260,JPY,0,0
-280_00,280,JPY,0,0
-290_00,290,JPY,0,0
-300_00,300,JPY,0,0
-310_00,310,JPY,0,0
-320_00,320,JPY,0,0
-330_00,330,JPY,0,0
-340_00,340,JPY,0,0
-350_00,350,JPY,0,0
-360_00,360,JPY,0,0
-370_00,370,JPY,0,0
-380_00,380,JPY,0,0
-390_00,390,JPY,0,0
-400_00,400,JPY,0,0
-410_00,410,JPY,0,0
-420_00,420,JPY,0,0
-430_00,430,JPY,0,0
-440_00,440,JPY,0,0
-450_00,450,JPY,0,0
-460_00,460,JPY,0,0
-470_00,470,JPY,0,0
-480_00,480,JPY,0,0
-490_00,490,JPY,0,0
-500_00,500,JPY,0,0
-510_00,510,JPY,0,0
-520_00,520,JPY,0,0
-530_00,530,JPY,0,0
-540_00,540,JPY,0,0
-550_00,550,JPY,0,0
-560_00,560,JPY,0,0
-570_00,570,JPY,0,0
-580_00,580,JPY,0,0
-590_00,590,JPY,0,0
-600_00,600,JPY,0,0
-610_00,610,JPY,0,0
-620_00,620,JPY,0,0
-630_00,630,JPY,0,0
-640_00,640,JPY,0,0
-650_00,650,JPY,0,0
-660_00,660,JPY,0,0
-670_00,670,JPY,0,0
-680_00,680,JPY,0,0
-690_00,690,JPY,0,0
-700_00,700,JPY,0,0
-710_00,710,JPY,0,0
-720_00,720,JPY,0,0
-730_00,730,JPY,0,0
-740_00,740,JPY,0,0
-750_00,750,JPY,0,0
-760_00,760,JPY,0,0
-770_00,770,JPY,0,0
-780_00,780,JPY,0,0
-790_00,790,JPY,0,0
-800_00,800,JPY,0,0
-810_00,810,JPY,0,0
-830_00,830,JPY,0,0
-840_00,840,JPY,0,0
-850_00,850,JPY,0,0
-860_00,860,JPY,0,0
-870_00,870,JPY,0,0
-880_00,880,JPY,0,0
-900_00,900,JPY,0,0
-910_00,910,JPY,0,0
-930_00,930,JPY,0,0
-940_00,940,JPY,0,0
-950_00,950,JPY,0,0
-960_00,960,JPY,0,0
-970_00,970,JPY,0,0
-980_00,980,JPY,0,0
-990_00,990,JPY,0,0
-1000_00,1000,JPY,0,0
-1010_00,1010,JPY,0,0
-1020_00,1020,JPY,0,0
-1030_00,1030,JPY,0,0
-1040_00,1040,JPY,0,0
-1050_00,1050,JPY,0,0
-1090_00,1090,JPY,0,0
-1100_00,1100,JPY,0,0
-1140_00,1140,JPY,0,0
-1150_00,1150,JPY,0,0
-1200_00,1200,JPY,0,0
-1240_00,1240,JPY,0,0
-1250_00,1250,JPY,0,0
-1300_00,1300,JPY,0,0
-1310_00,1310,JPY,0,0
-1360_00,1360,JPY,0,0
-1420_00,1420,JPY,0,0
-1470_00,1470,JPY,0,0
-1520_00,1520,JPY,0,0
-1570_00,1570,JPY,0,0
-1620_00,1620,JPY,0,0
-1680_00,1680,JPY,0,0
