--- gtfs_parser/tests/fixture/office_jp.txt
@@ -1,5 +1,0 @@
-﻿office_id,office_name,office_phone
-1000,帯広営業所,0155-31-8811
-2000,鹿追営業所,0156-66-2224
-3000,新得営業所,0156-64-5202
-4000,上士幌営業所,01564-2-2029
