--- gtfs_parser/tests/fixture/routes.txt
@@ -1,33 +1,0 @@
-﻿route_id,agency_id,route_short_name,route_long_name,route_desc,route_type,route_color,route_text_color
-12_A,8460101001629,12,一中・療養所線,帯広駅～国立病院循環（左回り・一中先回り・西5条3丁目経由）,3,FF0000,FFFFFF
-11_B,8460101001629,11,一中・療養所線,帯広駅～国立病院循環（右回り・協立先回り・西5条3丁目経由）,3,FF0000,FFFFFF
-16_F,8460101001629,16,一中線,帯広駅～拓殖本社（一中前経由）,3,FF0000,FFFFFF
-21_A,8460101001629,21,南商業高校線,帯広駅～南商高循環（第一病院・自由が丘団地経由）,3,800080,FFFFFF
-31_A,8460101001629,31,雄飛が丘団地線,帯広駅～雄飛が丘北区,3,008000,FFFFFF
-32_B,8460101001629,32,音更高校線(雄飛が丘経由),帯広駅～音更高校前（雄飛が丘経由）,3,008000,FFFFFF
-41_A,8460101001629,41,中鈴蘭線,帯広駅～中鈴蘭循環（左回り・十勝大橋先回り）,3,FF8000,000000
-42_B,8460101001629,42,中鈴蘭線,帯広駅～中鈴蘭循環（右回り・すずらん大橋先回り）,3,FF8000,000000
-51_C,8460101001629,51,然別湖線,帯広駅～然別湖,3,0000FF,FFFFFF
-53_F,8460101001629,53,新得帯広線(新帯線),帯広駅～新得,3,0000FF,FFFFFF
-52_D,8460101001629,52,鹿追線,帯広駅～鹿追,3,0000FF,FFFFFF
-61_A,8460101001629,61,上士幌線,帯広駅～上士幌,3,FF80FF,000000
-62_B,8460101001629,62,音更高校上士幌線(音上線),帯広駅～上士幌（音更高校経由）,3,FF80FF,000000
-33_C,8460101001629,33,緑陽台駒場線(緑駒線),帯広駅～音更高・駒場循環（緑陽台・音更大通経由）,3,008000,FFFFFF
-24_C,8460101001629,24,中鈴蘭・南商線,おおや整形～帯広駅～南商業高校（木野農協経由）,3,FF8000,000000
-25_C,8460101001629,25,帯広の森線,帯広駅～帯広の森・白樺学園（市役所南口・交流館・南商経由）,3,800080,FFFFFF
-22_B,8460101001629,22,南商業高校線,南商業高校～帯広駅（自由が丘団地・市役所南口経由）,3,800080,FFFFFF
-15_E,8460101001629,15,療養所線,【公園発】帯広駅～拓殖本社（西5条3丁目・協立病院経由）,3,FF0000,FFFFFF
-14_C,8460101001629,14,一中・療養所線,帯広駅～国立病院循環（左回り・一中先回り・市役所南口経由）,3,FF0000,FFFFFF
-13_D,8460101001629,13,一中・療養所線,帯広駅～国立病院循環（右回り・協立先回り・市役所南口経由）,3,FF0000,FFFFFF
-23_G,8460101001629,23,一中・南商線,拓殖本社～帯広駅～南商業高校（一中前経由）,3,FF0000,FFFFFF
-17_H,8460101001629,17,運転免許試験場線,帯広駅～免許試験場（西2条9・西5条3・協立病院経由）,3,FF0000,FFFFFF
-53-2_G,8460101001629,53-2,新得鹿追線(新鹿線),鹿追～新得駅前,3,0000FF,FFFFFF
-26_D,8460101001629,26,帯広の森線,白樺学園・帯広の森～帯広駅（自由が丘団地・第一病院経由）,3,800080,FFFFFF
-34_A,8460101001629,34,音更高校線(国道経由),帯広駅～音更高校（国道経由）,3,0000FF,FFFFFF
-54_E,8460101001629,54,柏林台音更高校スクール線,柏林台駅～音更高校（スクール）,3,0000FF,FFFFFF
-71_H,8460101001629,71,屈足線,幸町4丁目～南新得,3,0000FF,FFFFFF
-音更町コミュニティバス線_A,8460101001629,,音更コミバス(すずらん号),すずらん号 木野コース・音更コース,3,800040,FFFFFF
-新得町コミュニティバス線_A,8460101001629,,新得コミバス(新得市街地線),新得市街地 西地区・東地区線（西地区：フクハラ経由）,3,FF80FF,000000
-新得町コミュニティバス線_B,8460101001629,,新得コミバス(屈足地区線),屈足地区線（新得駅～屈足～新得駅）,3,FF80FF,000000
-新得町コミュニティバス線_C,8460101001629,,新得コミバス(屈足早朝線),屈足地区線・早朝便,3,FF80FF,000000
-清水町コミュニティバス線_A,8460101001629,,清水コミバス,西地区～東地区～御影～清水【月・水・金運行】,3,0000A0,FFFFFF
